Create GridSearch_RF_Parallel
In my needing to do a grid search for a random forest, differently from the method rf in the caret package, I build this structure to do a 5-fold cross validation grid search in the ntree, nodesize pair. Its is easy to add some others parameters.

Also I use the doParallel package to make it faster.... it is my first attempt to parallel programming, so I'm sorry if i did something not optimal.